made useOpen hook
add  useOpen hook for ContactsAll and CustomSelect
