+++ src/Components/Header/Desktop/ContactsAll.jsx
@@ -1,37 +1,18 @@
-import React, { useRef, useState, useEffect } from 'react';
+import React from 'react';
+import { useOpen } from '../../hooks/useOpen';
 
 export default function ContactsAll() {
-    const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-    const dropdownRef = useRef();
-    const ref = useRef();
-
-    function toggleDropdown() {
-        setIsDropdownOpen(!isDropdownOpen);
-    }
-
-    function closingClickOutside(event) {
-        if (dropdownRef.current && !ref.current.contains(event.target)) {
-            setIsDropdownOpen(false);
-        }
-    };
-
-    useEffect(() => {
-        document.addEventListener('mousedown', closingClickOutside);
-        return () => {
-            document.removeEventListener('mousedown', closingClickOutside);
-        };
-    }, [isDropdownOpen]);
-
+    const { open, setOpen, ref, dropRef, toggleDrop } = useOpen(); //функционал по открытию-закрытию drop
 
     return (
         <div ref={ref} className='contacts__all'>
             <div className='contacts__dropdown dropdown'>
-                <a className="dropdown__toggle" href="tel:111" onClick={toggleDropdown} aria-expanded={isDropdownOpen}>
+                <a className="dropdown__toggle" href="tel:111" onClick={toggleDrop} aria-expanded={open}>
                     111 (A1, МТС, Life)
                     <br />
                     +375 (17) 111-11-11
                 </a>
-                <div ref={dropdownRef} className={`dropdown__menu ${isDropdownOpen ? 'show' : ''}`}>
+                <div ref={dropRef} className={`dropdown__menu ${open ? 'show' : ''}`}>
                     <div className="contacts__list">
                         <div className="contacts__item"><a href="tel:111">111 (A1, МТС, Life)</a></div>
                         <div className="contacts__item"><a href="tel:+375291121111"><span>+375 (29)</span> 112 11 11</a></div>
